feat(pvmt): Rework the PVMT extension [WIP]
The PVMT model extension (`capellambse.extensions.pvmt`) used to use a
parallel package to `capellambse.model` for accessing and manipulating
managed property values (`capellambse.pvmt`). This approach had several
shortcomings, which manifested in a very different API design and
missing functionality that is now commonplace in the rest of the model.

With this refactor, the PVMT model extension has been rewritten from
scratch on top of the `capellambse.model` package. This addresses the
aforementioned shortcomings and vastly improves and extends the API,
making it more consistent and allowing to leverage new functionality.

The legacy `capellambse.pvmt` library has been deprecated and will be
removed in v0.6.